% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.chebyshev_dist}
\alias{.chebyshev_dist}
\title{Compute pairwise Chebyshev (L-infinity) distance matrix}
\usage{
.chebyshev_dist(mat)
}
\arguments{
\item{mat}{Numeric matrix (rows = observations).}
}
\value{
Symmetric distance matrix.
}
\description{
Compute pairwise Chebyshev (L-infinity) distance matrix
}
