% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.discretize}
\alias{.discretize}
\title{Discretize a numeric vector into integer bin labels}
\usage{
.discretize(x, bins = 10)
}
\arguments{
\item{x}{Numeric vector}

\item{bins}{Number of bins}
}
\value{
Integer vector of bin indices (1..bins)
}
\description{
Discretize a numeric vector into integer bin labels
}
