% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.scale_unit}
\alias{.scale_unit}
\title{Scale vector to [0, 1]}
\usage{
.scale_unit(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector scaled to [0, 1]
}
\description{
Scale vector to [0, 1]
}
