% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartMI-package.R
\docType{package}
\name{PartMI-package}
\alias{PartMI}
\alias{PartMI-package}
\title{PartMI: Part Mutual Information for Direct Association Network Inference}
\description{
Implements Part Mutual Information (PMI) for inferring direct associations
in biological networks. PMI extends standard Mutual Information (MI) by
conditioning on background variables, removing indirect dependencies and
reducing false-positive edges in network reconstruction.
}
\section{Core functions}{

\itemize{
\item \code{\link{mi}} - Compute mutual information between two variables
\item \code{\link{pmi}} - Compute part (conditional) mutual information
\item \code{\link{pmi_network}} - Infer direct association network
\item \code{\link{normalize_data}} - Normalize expression data
\item \code{\link{discretize_data}} - Discretize continuous data
}
}

\references{
Zhao et al. (2016) "Part mutual information for quantifying direct
associations in networks." PNAS 113(18): 5130-5135.

Frenzel & Pompe (2007) "Partial Mutual Information for Coupling Analysis
of Multivariate Time Series." Physical Review Letters 99(20): 204101.

Kraskov, Stögbauer & Grassberger (2004) "Estimating mutual information."
Physical Review E 69(6): 066138.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/danymukesha/PartMI}
  \item Report bugs at \url{https://github.com/danymukesha/PartMI/issues}
}

}
\author{
\strong{Maintainer}: Dany Mukesha \email{danymukesha@gmail.com} (\href{https://orcid.org/0009-0001-9514-751X}{ORCID})

Other contributors:
\itemize{
  \item Zhao et al. (Original PMI method) [contributor]
}

}
\keyword{internal}
