% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{print.pmi_network}
\alias{print.pmi_network}
\title{Print method for pmi_network objects}
\usage{
\method{print}{pmi_network}(x, ...)
}
\arguments{
\item{x}{A \code{pmi_network} object.}

\item{...}{Additional arguments.}
}
\description{
Print method for pmi_network objects
}
